+++ index.html
@@ -11,7 +11,7 @@
   <link rel="stylesheet" href="http://fonts.googleapis.com/css?family=Roboto:300,400,500,700" type="text/css">
 
   <script defer src="https://code.getmdl.io/1.3.0/material.min.js"></script>
-  <link rel="stylesheet" href="assets/team.css">
+  <link rel="stylesheet" href="assets/style/team.css">
   <title>Cohort Collab</title>
 </head>
 
@@ -70,12 +70,9 @@
       <p class="footerParagraph">2019 &copy; Cohort Collab | All Rights Reserved.</p>
     </div>
   </footer>
-
-
-
    
   <script src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
-  <script src="assets/team.js"></script>
+  <script src="assets/js/index.js"></script>
 </body>
 
 </html>
